+++ Assets/Input/Inputs.cs
@@ -1,3 +1,4 @@
+using Interaction;
 using UnityEngine;
 using UnityEngine.InputSystem;
