--- Assets/Scripts/AnimationCotrollers/CatapultController.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.InputSystem;
-using static UnityEngine.Rendering.DebugUI;
-
-public class CatapultController : MonoBehaviour
-{
-    // DEPRECATED
-    // This Functionality was moved to Inputs.cs
-}
