--- Assets/Scripts/AnimationCotrollers/CatapultController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 84bc81cbcd179634da22916454335463
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
