--- Assets/Scripts/AnimationCotrollers/CrossbowmanAnimationController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f8f3862d06379b64b82ebed27fc68689
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
