+++ Assets/Scripts/Enemies/Footsoldier.cs.meta
