--- Assets/Scripts/InteractableCatapult.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using Unity.VisualScripting;
-using UnityEngine;
-using Random = UnityEngine.Random;
-
-[Tooltip("DEPRECATED")]
-public class InteractableCatapult : MonoBehaviour
-{
-    // This functionality was moved to the Catapult.cs Script.
-}
