--- Assets/Scripts/InteractableCatapult.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5f3e38c3b1982434faad7317e6fa4441
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
