+++ Assets/Scripts/Interaction.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4465da24da70d3f4b8e04f4c5d3185c5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
